refactor: Split Bun lockfile module
